requirements.txt
aiohttp>=3.9.0
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
httpx>=0.27.0
apscheduler>=3.10.0
